Added DTO for Author

+++ src/main/java/com/sumerge/course_recommender/course/CourseService.java
@@ -1,6 +1,7 @@
 package com.sumerge.course_recommender.course;
 
 import com.sumerge.course_recommender.mapper.MapStructMapper;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import java.util.UUID;
 
 @Service//@Component
+@Transactional
 public class CourseService {
 
     private final CourseRecommender courseRecommender;
@@ -45,7 +47,12 @@ public class CourseService {
     }
 
     public void deleteCourse(UUID course_id){
-        courseRepository.deleteById(course_id);
+        if (courseRepository.existsById(course_id)){
+            courseRepository.deleteById(course_id);
+        }
+//        else{
+//            throw
+//        }
     }
 
 }
